refactor: update error codes and integrate IGDB and IMGBB services
- Renamed error code GAME_SERVICE_UNAVAILABLE to IGDB_SERVICE_UNAVAILABLE.
- Introduced CacheKeys interface in cache service for better cache management.
- Refactored UploadService to utilize IntegrationsService for image uploads.
- Added HardcoverService, IGDBService, IMGBBService, JikanService, and TMDBService with basic structure and tests.
- Implemented IGDBService with methods for obtaining access tokens and searching games.
- Created IMGBBService for handling image uploads.
- Established IntegrationsModule to encapsulate all integration services.
- Added unit tests for all new services to ensure basic functionality.
