+++ src/shared/constants/error-codes.ts
@@ -28,7 +28,7 @@ export const ERROR_CODES = {
 	},
 	FAILED_TO_UPLOAD_IMAGE: { message: "FAILED_TO_UPLOAD_IMAGE", status: 400 },
 	RATE_LIMIT_EXCEEDED: { message: "RATE_LIMIT_EXCEEDED", status: 429 },
-	GAME_SERVICE_UNAVAILABLE: {
+	IGDB_SERVICE_UNAVAILABLE: {
 		message: "GAME_SERVICE_UNAVAILABLE",
 		status: 503,
 	},
