+++ src/shared/infra/integrations/hardcover.service.spec.ts
@@ -1,0 +1,5 @@
+describe("HardcoverService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
