+++ src/shared/infra/integrations/hardcover.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class HardcoverService {}
