+++ src/shared/infra/integrations/igdb.service.spec.ts
@@ -1,0 +1,5 @@
+describe("IGDBService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
