+++ src/shared/infra/integrations/imgbb.service.spec.ts
@@ -1,0 +1,5 @@
+describe("IMGBBService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
