+++ src/shared/infra/integrations/jikan.service.spec.ts
@@ -1,0 +1,5 @@
+describe("JikanService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
