+++ src/shared/infra/integrations/tmdb.service.spec.ts
@@ -1,0 +1,5 @@
+describe("TMDBService", () => {
+	it("true is true", () => {
+		expect(true).toBe(true);
+	});
+});
