+++ src/shared/infra/upload/upload.service.ts
@@ -1,18 +1,13 @@
-import { HttpService } from "@nestjs/axios";
-import { Injectable, Logger } from "@nestjs/common";
-import { ConfigService } from "@nestjs/config";
-import { firstValueFrom } from "rxjs";
-import { ERROR_CODES } from "@/shared/constants/error-codes";
-import { AppException } from "@/shared/exceptions/app.exceptions";
+import { Injectable } from "@nestjs/common";
+import { HttpService } from '@nestjs/axios';
+import { firstValueFrom } from 'rxjs';
+
+import { IntegrationsService } from '../integrations/integrations.service';
 
 @Injectable()
 export class UploadService {
-	private readonly logger = new Logger(UploadService.name);
-
-	private readonly IMGBB_API_URL = "https://api.imgbb.com/1";
-
 	constructor(
-		private readonly configService: ConfigService,
+		private readonly integrationsService: IntegrationsService,
 		private readonly httpService: HttpService,
 	) {}
 
@@ -25,28 +20,10 @@ export class UploadService {
 
 		const buffer = Buffer.from(response.data, "binary");
 
-		return this.uploadFromBuffer(buffer);
+		return this.integrationsService.imgbb.upload(buffer);
 	}
 
 	async uploadFromBuffer(buffer: Buffer): Promise<string> {
-		try {
-			const formData = new FormData();
-
-			formData.append("image", buffer.toString("base64"));
-
-			const response = await firstValueFrom(
-				this.httpService.post(`${this.IMGBB_API_URL}/upload`, formData, {
-					params: {
-						key: this.configService.get<string>("IMGBB_API_KEY"),
-					},
-				}),
-			);
-
-			return response.data?.data?.image?.url as string;
-		} catch (error) {
-			this.logger.error("Failed to upload image to ImgBB", error);
-
-			throw new AppException(ERROR_CODES.FAILED_TO_UPLOAD_IMAGE);
-		}
+		return this.integrationsService.imgbb.upload(buffer);
 	}
 }
